Fix warnings (and a few bugs) Remove dead code
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@23439 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ db/drivers/odbc/create_table.c
@@ -28,7 +28,7 @@ db__driver_create_table (dbTable *table)
 
     if ((ret != SQL_SUCCESS) && (ret != SQL_SUCCESS_WITH_INFO)) {
 	SQLGetDiagRec(SQL_HANDLE_STMT, c->stmt, 1, NULL, &err, msg, sizeof(msg), NULL);
-	G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", db_get_string(&sql), msg, err);
+	G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", db_get_string(&sql), msg, (int) err);
 	report_error(emsg);
         G_free(emsg);
