+++ db/drivers/odbc/execute.c
@@ -26,7 +26,7 @@ db__driver_execute_immediate (dbString *sql)
     if ((ret != SQL_SUCCESS) && (ret != SQL_SUCCESS_WITH_INFO))
     {
 	SQLGetDiagRec(SQL_HANDLE_STMT, c->stmt, 1, NULL, &err, msg, sizeof(msg), NULL);
-	G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", s, msg, err);
+	G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", s, msg, (int) err);
 	report_error(emsg);
         G_free(emsg);
