+++ db/drivers/odbc/table.c
@@ -166,7 +166,7 @@ dbString *name;
     if ((ret != SQL_SUCCESS) && (ret != SQL_SUCCESS_WITH_INFO))
     {
         SQLGetDiagRec(SQL_HANDLE_STMT, c->stmt,1, NULL, &err,msg,sizeof(msg),NULL);
-        G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", cmd, msg, err);
+        G_asprintf(&emsg, "SQLExecDirect():\n%s\n%s (%d)\n", cmd, msg, (int) err);
         report_error(emsg);
         G_free(emsg);
